# Clear effect substitutions when exiting toplevel opdefs